+++ include/view/view.h
@@ -13,7 +13,10 @@ void view_fatal(
 extern PyObject* ip_address;
 extern PyObject* invalid_status_error;
 extern PyObject* route_log;
+extern PyObject* route_warn;
 extern PyObject* ws_cls;
+extern PyObject* ws_disconnect_err;
+extern PyObject* ws_err_cls;
 
 #if defined(__LINE__) && defined(__FILE__)
 #define VIEW_FATAL(msg) view_fatal(msg, __FILE__, __func__, __LINE__)