+++ src/view/_util.py
@@ -57,7 +57,9 @@ class LoadChecker:
     _view_loaded: bool
 
     def _view_load_check(self) -> None:
-        if (not self._view_loaded) and (not os.environ.get("_VIEW_CANCEL_FINALIZERS")):
+        if (not self._view_loaded) and (
+            not os.environ.get("_VIEW_CANCEL_FINALIZERS")
+        ):
             warnings.warn(f"{self} was never loaded", NotLoadedWarning)
 
     def __post_init__(self) -> None:
@@ -74,7 +76,9 @@ def shell_hint(*commands: str) -> Panel:
     if os.name == "nt":
         shell_prefix = f"{os.getcwd()}>"
     else:
-        shell_prefix = f"{getpass.getuser()}@{socket.gethostname()}[bold green]$[/]"
+        shell_prefix = (
+            f"{getpass.getuser()}@{socket.gethostname()}[bold green]$[/]"
+        )
 
     formatted = [f"{shell_prefix} {escape(command)}" for command in commands]
     return Panel.fit(